Normalize v4 feature flags to v3 format.
This ensures backwards compatibility.

Reorganize tests and add v4 tests.
Move the local eval tests into their own file.
Move feature flag tests into `FeatureFlagTest.php`
