+++ test/PostHogTest.php
@@ -252,117 +252,6 @@ class PostHogTest extends TestCase
         );
     }
 
-    public function testIsFeatureEnabled()
-    {
-        $this->assertFalse(PostHog::isFeatureEnabled('having_fun', 'user-id'));
-        $this->assertEquals(
-            $this->http_client->calls,
-            array(
-                0 => array(
-                    "path" => "/api/feature_flag/local_evaluation?send_cohorts&token=random_key",
-                    "payload" => null,
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION, 1 => 'Authorization: Bearer test'),
-                    "requestOptions" => array(),
-                ),
-                1 => array(
-                    "path" => "/decide/?v=3",
-                    "payload" => sprintf('{"api_key":"%s","distinct_id":"user-id","person_properties":{"distinct_id":"user-id"}}', self::FAKE_API_KEY),
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION),
-                    "requestOptions" => array("timeout" => 3000, "shouldRetry" => false),
-                ),
-            )
-        );
-    }
-
-    public function testIsFeatureEnabledGroups()
-    {
-        $this->assertFalse(PostHog::isFeatureEnabled('having_fun', 'user-id', array("company" => "id:5")));
-
-        $this->assertEquals(
-            $this->http_client->calls,
-            array(
-                0 => array(
-                    "path" => "/api/feature_flag/local_evaluation?send_cohorts&token=random_key",
-                    "payload" => null,
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION, 1 => 'Authorization: Bearer test'),
-                    "requestOptions" => array(),
-                ),
-                1 => array(
-                    "path" => "/decide/?v=3",
-                    "payload" => sprintf(
-                        '{"api_key":"%s","distinct_id":"user-id","groups":{"company":"id:5"},"person_properties":{"distinct_id":"user-id"},"group_properties":{"company":{"$group_key":"id:5"}}}',
-                        self::FAKE_API_KEY
-                    ),
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION),
-                    "requestOptions" => array("timeout" => 3000, "shouldRetry" => false),
-                ),
-            )
-        );
-    }
-
-    public function testGetFeatureFlag()
-    {
-        $this->assertEquals("variant-value", PostHog::getFeatureFlag('multivariate-test', 'user-id'));
-        $this->assertEquals(
-            $this->http_client->calls,
-            array(
-                0 => array(
-                    "path" => "/api/feature_flag/local_evaluation?send_cohorts&token=random_key",
-                    "payload" => null,
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION, 1 => 'Authorization: Bearer test'),
-                    "requestOptions" => array(),
-                ),
-                1 => array(
-                    "path" => "/decide/?v=3",
-                    "payload" => sprintf('{"api_key":"%s","distinct_id":"user-id","person_properties":{"distinct_id":"user-id"}}', self::FAKE_API_KEY),
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION),
-                    "requestOptions" => array("timeout" => 3000, "shouldRetry" => false),
-                ),
-            )
-        );
-    }
-
-    public function testGetFeatureFlagDefault()
-    {
-        $this->assertEquals(PostHog::getFeatureFlag('blah', 'user-id'), null);
-
-        $this->checkEmptyErrorLogs();
-    }
-
-    public function testGetFeatureFlagGroups()
-    {
-        $this->assertEquals(
-            "variant-value",
-            PostHog::getFeatureFlag('multivariate-test', 'user-id', array("company" => "id:5"))
-        );
-
-        $this->assertEquals(
-            $this->http_client->calls,
-            array(
-                0 => array(
-                    "path" => "/api/feature_flag/local_evaluation?send_cohorts&token=random_key",
-                    "payload" => null,
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION, 1 => 'Authorization: Bearer test'),
-                    "requestOptions" => array(),
-                ),
-                1 => array(
-                    "path" => "/decide/?v=3",
-                    "payload" => sprintf(
-                        '{"api_key":"%s","distinct_id":"user-id","groups":{"company":"id:5"},"person_properties":{"distinct_id":"user-id"},"group_properties":{"company":{"$group_key":"id:5"}}}',
-                        self::FAKE_API_KEY
-                    ),
-                    "extraHeaders" => array(0 => 'User-Agent: posthog-php/' . PostHog::VERSION),
-                    "requestOptions" => array("timeout" => 3000, "shouldRetry" => false),
-                ),
-            )
-        );
-    }
-
-    public function testfetchFeatureVariants()
-    {
-        $this->assertIsArray(PostHog::fetchFeatureVariants('user-id'));
-    }
-
     public function testEmptyProperties(): void
     {
         self::assertTrue(
